proyecto/requirements.txt
asgiref==3.4.1
Django==3.2.8
Pillow==8.3.2
pkg-resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
